amend! amend! Caused a line to be rendered!
Caused a line to be rendered!

Turns out many things were wrong with that commit. Here's a fix to the
first one:
Fixed orientation of x and y such that the x is now the horizontal and y
is the vertical, as intended.

And the second:
Fixed the lerp code, following Freya Holmer's video on Bezier Curves.
Also added another line to demo this change.