Got bezier curves working!
They look absolutely disgusting, and only one has been tested, but I'm
so glad they worked first try!
I bet the reason they're so low-res has something to do with how much
I'm trying to do with integral types, and not floats... that's certainly
something to try.
But this was a great way to spend a sick afternoon! I did this with no
tutorial! Yay me!