Removed the note that vec.rs was evil.
I don't think it is anymore. I'm not sure if it was ever committed in
the evil state. I don't even know if the evil state compiled!
I think the evil state was trying to define operators on tuples of
integral types. Likely the orphan rules prevented this evilness. This is
all speculation, though.