Tried to make a zero-cost generic bezier function.
It didn't really work, for understandable reasons (attempt to use a
non-const expression in a const expression context), even though it
'should' be something I *think* could work if I could force loops to be
statically unrolled. It might be worth pursuing this while accepting
some cost (e.g. a vec) or if I define a loop macro of some sort that
can be unrolled at compile time.