Now its an SDL2 application, hoorah.
This took a long time and also seems to have necessitated unsafe??? What
the heck?
I dunno man, but it does feel very powerful to get it working.
I don't like SDL2 all that much right now, but I do understand it okay
at the moment, and I'm loathe to learn a new thing. It does seem like
it's a pretty industry-standard thing, at least in the world of gamedev.
But I should also consider:
- miniquad, if I want something in Rust that is cross-platform
  (including maybe iOS?)
- pix-win-loop, if I want something in Rust that's very quick and easy
  and doesn't allow any GPU anything
- I dunno man, anything else.

I can already feel the Stockholm syndrome setting in, though... it does
feel really good to get this working, and at such a low-level! If it's
this hard, it *must* be worthwhile!

It wasn't even that hard, I was just slow...