Added a dialog that lets you save the image as a bmp.
It's kinda funny to me that this isn't a feature of SDL2, but rather
something I have to add with another crate. But hey, it's kinda cool.

Something about GUI programming feels more real, even when it is this
freaking simple.

Preview doesn't like this .bmp, but VSCode does! Or whatever extension
is currently showing it in VSCode.