Switched to 60 fps!
Also everything except saving to a .bmp works on my install of Asahi
Linux! Just had to install a few packages, including SDL2-devel to
dynamically link, and cmake and libXext-devel to statically link. cmake
may have been necessary for all of the sorts of linking. The versions
from the default repositories were sufficient, thankfully!
When running the dynamically linked version, it appears use Wayland, and
when running the statically linked version, it appears to use X. Wayland
respects my resolution scaling but shows spurious handle affordances
when hovering around the edge of the window. Wayland has *much* nicer
"message boxes".
It's *so* cool that this works at all, let alone as well as it does.