Enabled static-link option so CodeLLDB works.
Otherwise, it would fail to launch the process - for that matter,
manually invoking the built executable would fail, too - I believe with
the same message.

Also, if I step over lines in the debugger, it appears to unfreeze the
other threads! So I can watch the image get drawn, just like I wanted!