Added interior mutability just so if you quit the app before selecting a world, the app does quit when you exit that file selector.
Also ran cargo fmt.
I hope this isn't necessary, but I'm not sure yet. Good practice, I
guess!
Interior mutability is not very intuitive. I still think it's kinda
elegant to have just *one* way that the rules get broken, though it's
hard to explain it (I'm talking, of course, about this as an analogue to
C++'s mutable keyword for class members, though it's much more prevalent
here because of aliasing rules than it was in C++).
I want to believe the trickiness here was actually just having to do
with sharing something between threads, but I'm not sure.
It was just so much craziness that I forgot what I was doing for a bit
in the middle.
I wonder what Mojo's attitude re: Interior Mutability and thread-safety
are...

Anyway, this does at least add a feature, underwhelming though it is -
your wish to quit the app is respected sooner.
Unfortunately, I'm fairly certain it can't actually be respected
*immediately* unless I were to introduce an async runtime. It is good
that rfd (rusty file dialogs) supports creating dialogs asynchronously,
but I just *don't* want to deal with async right now. I hardly
understand it at the best of times.