Changed control-flow flag variable save_file to save_bmp_image.
Even if it shouldn't need a name, at least its name can be descriptive.
(I am staunchly opposed to control-flow-flag variables in principle).
This change was also bundled with another change on a branch whose name
was, at least originally, a sort of plea about raw_window_handle. The
branch's name is i-really-want-raw-window-handle-to-be-good-but-its-not-being-nice-to-me
as of writing this commit, and *that* commit is commit
b552dfd65787bc3e628f713fa259cda11e326e7b.