Canceling in file dialogs now cancels the action.
Also, changed error handling such that terminations that report an error
`panic!` immediately and terminations that don't cause the main function
to return `Ok(())`. This is to ensure destructors run in all cases (this
code is not intended to be compiled such that panics abort, and is
currently configured to build such that they unwind, which is the
default).
Documented control-flow weirdness with how saving an image is handled in
comments on main.rs.