Unnecessary bit of code I don't want to forget.
I wrote this as a way for a function to indicate that the program should
terminate normally, presumably by returning Ok(()) or () from it's main
function. Basically, it was a way to unwind the stack without panicking.
It is better documented in a doc comment in its file.

I ended up not needing this because, currently, no function other than
main indicates that the program should end. This feels cleaner.

I'm recording this mostly because it's the first time I tried to write a
general abstraction for "error" handling. I remember struggling with
this same problem back in CS 141, in our final program, which I believe
was an arbitrary-length-integer calculator. I eschewed exceptions then,
much as I eschew trickery here.