amend! Made bundling and static-linking optional features.
Made bundling and static-linking optional features.

"bundled" is required for my setup on macOS and Windows, but it should
build with no features specified on Fedora Asahi Linux with the sdl2
development library installed (`SDL2-devel`).

I will update with an amendment commit if this works.

amendment:

It works! I also changed the default features and the
launch.json so they bundle by default, so both debugging with vscode and
running with `cargo run` "just work" on all my platforms.