Factored out the ugliest part of the UnsafeCell code.
It makes sense that as_ref() on an arbitrary pointer might fail, but I don't think
it will fail when the pointer is gotten through UnsafeCell::get(). So we
really do just need to have two unwrap_unchecked()'s for maximum
performance.

Again, this seems to work, but take that with a giant heaping of salt.