amend! Made it work using the atomic crate!
Made it work using the atomic crate!

Turns out you have to make sure the type is aligned to 4, or it requires
the locking fallback, and will not work with native atomic instructions.
This code also prints a message saying whether or not the color uses any
locks to the console, though in its current state it shouldn't even
compile if it would need to use a lock, since the atomic crate's
"fallback" feature is disabled.