Validates the length of a spline, now!
Following the advice on this pull request:
https://github.com/serde-rs/serde/pull/1858#pullrequestreview-575089757
This is arguably an abuse of the deserialization system - it really
ought to express how to *parse*, not validate.

It might be better to do this by creating a newtype that expresses the
desired constraint. This sort of unifies "parsing" and "validation" in
my mind, and it provides some stronger guarantees about what a valid
instance of that type is - presuming you write a more limiting interface
for that type, that is.