Ran cargo clippy.
This was my first time.
I was overall pleased with the recommendations, and I find it invaluable
that the lints are maintained by the same people who write the language.
If this is what we owe Go, I think we owe Go a lot. Of course, we also
owe the hardworking developers who wrote *these* lints.
I find it almost baffling that it seems like there is no built-in way to
cargo to make running "cargo clippy" something that happens whenever I
build my project. Of course, I should be careful what I wish for - my
favorite thing about cargo is that it is relatively simple (especially
compared to maven), and even allowing cargo to define it's own aliases
is a step towards suspect complexity (very few ways you could invoke
cargo would lack the ability to define aliases, though it is nice to
have a unified way to define them).
As for the lints themselves:
I agreed with every one reported here, and the recommendation for how to
use the fancy functional verbs was very appreciated, and reminiscent of
the lints IntelliJ gave me in my recent Concurrent Programming class.
Nothing leapt out at me like my first automatic Java lints did, though -
no detection of unchanging loop conditions or anything like that (I
didn't test anything, though). But all in all, it's very pleasant that
the recommendations from the rust tooling seem to continue to be
consistent and helpful.